% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_checkStyle.R
\name{checkStyle}
\alias{checkStyle}
\title{Check code style}
\usage{
checkStyle(files, type = c("script", "pkgFuns"))
}
\arguments{
\item{files}{character vector: One or more filepaths}

\item{type}{character: Type of the file}
}
\description{
Checks a file for violations of the INWT style conventions using
\code{\link[lintr]{lint}}. In addition to the \code{\link[lintr]{linters}}
provided by the package \code{lintr}, some custom linters are tested. The set
of used linters depends on the type of the file. If the type is specified
("script" or "pkgFuns"), additional linters are tested. For details about the
tested linters see \code{\link{selectLntrs}}.
}
\examples{
\dontrun{
writeLines(con = "lintExample.txt",
           text = c("# Example script to demonstrate INWT's own linters",
           # nolint start
                    "",
                    "foo <- function(x = 1, y) {",
                    "  2*x + 1",
                    "}",
                    "",
                    paste0("# This  line containts  double spaces and is ",
                           "very long. The following lines will use = ",
                           "instead of <- and access an internal INWT ",
                           "function."),
                    "z = 1",
                    "print(INWTUtils:::scriptLntrs())",
                    ""))
           # nolint end
checkStyle("lintExample.txt", type = "script")
unlink("lintExample.txt") # Remove file

}

}
