% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_createProjectSkeleton.R
\name{copyFile}
\alias{copyFile}
\title{Copy file from inst}
\usage{
copyFile(dir, origin, dest = origin, ...)
}
\arguments{
\item{dir}{character: project root}

\item{origin}{character: name/path of file relative to inst}

\item{dest}{character: new path (path to directory or path with filename)
relative to project root}
}
\description{
Copy a file from inst to the - per default - respective
directory relative to the project root
}
