% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_createProjectSkeleton.R
\name{createProject}
\alias{createProject}
\title{Create R project}
\usage{
createProject(pkg, pkgFolder = ".", dir = ".")
}
\arguments{
\item{pkg}{logical: Does the project contain a package?}

\item{pkgFolder}{character: Folder where the package lives. "." per
default, i.e., \code{dir}.}

\item{dir}{character: Directory where the R project is created; current
working directory by default}
}
\description{
Creates an R project with useful configuration. If the project
contains a package, the respective options are set. The project is named
after the folder which contains it.
Used in \code{\link{createProjectSkeleton}}.
}
\examples{
\dontrun{
createProject(pkg = FALSE, dir = "./")
dir.create("tmp")
createProject(pkg = TRUE, pkgOnToplevel = FALSE, dir = "tmp")}

}
