% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_createProjectSkeleton.R
\name{createProjectSkeleton}
\alias{createProjectSkeleton}
\title{Create file structure}
\usage{
createProjectSkeleton(dir = ".", pkgName = NULL, pkgOnToplevel = TRUE,
  rProject = FALSE, exampleScript = TRUE, ...)
}
\arguments{
\item{dir}{character: Directory where the file structure is created, relative
to the current working directory. The current working directory by default.}

\item{pkgName}{character: If \code{pkgName} is specified, a package with this
name is created.}

\item{pkgOnToplevel}{logical: Should the package live in the main directory
(default) or in a subfolder called package?}

\item{rProject}{logical: Create Rproject file?}

\item{exampleScript}{logical: Create example script? (not yet used)}

\item{...}{Further arguments passed to \code{\link[devtools]{create}} resp.
\code{\link[devtools]{setup}} if creating a package}
}
\description{
Creates a useful file and folder structure. This includes:
\itemize{
  \item data (folder)
  \item reports (folder)
  \item rScripts (folder)
  \item libWin (folder) containing .gitginore
  \item libLinux (folder) containing .gitginore
  \item .Rprofile
  \item .Rproj (optional; created with \code{\link{createProject}})
}
The infrastructure for a package can also be added optionally (created with
\code{\link{createPackage}}).
}
\examples{
createProjectSkeleton("tmp", rProject = TRUE)

}
