% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_rmPkgs.R
\name{deletePkgs}
\alias{deletePkgs}
\title{Uninstall package from all libraries}
\usage{
deletePkgs(packages)
}
\arguments{
\item{packages}{character vector of packages to remove}
}
\description{
Uninstalls one or more packages from all libraries with write
rights
}
