% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_formattingFigures.R
\name{formatBigNumbers}
\alias{formatBigNumbers}
\title{Format Big Numbers}
\usage{
formatBigNumbers(x, bigMark = "\\\\\\\\,", digits = 0)
}
\arguments{
\item{x}{numeric}

\item{bigMark}{character: thousand separator}

\item{digits}{integer}
}
\description{
Wrapper around \code{\link[base]{formatC}}
}
\examples{
formatBigNumbers(1234567.123, digits = 2)
formatBigNumbers(1234567, bigMark = " ")
}
