% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalLntrs.R
\name{INWTlinters}
\alias{INWTlinters}
\alias{args_without_default_first_linter}
\alias{double_space_linter}
\alias{internal_INWT_function_linter}
\alias{library_linter}
\alias{setwd_linter}
\alias{source_linter}
\title{INWT's own linters}
\usage{
INWTlinters()

args_without_default_first_linter(source_file)

double_space_linter(source_file)

internal_INWT_function_linter(source_file)

library_linter(source_file)

setwd_linter(source_file)

source_linter(source_file)
}
\arguments{
\item{source_file}{returned by \code{\link[lintr]{get_source_expressions}}}
}
\description{
Linters added by INWT. Can be used with
\code{\link[lintr]{lint}}.
}
\section{Functions}{
\itemize{
\item \code{args_without_default_first_linter}: Arguments without default values should come before
arguments with default values. (only for package functions)

\item \code{double_space_linter}: Are there double whitespaces ("  ")?

\item \code{internal_INWT_function_linter}: Instead of using internal functions from an INWT
package in scripts with \code{:::}, they should be exported and documented.
(only for scripts)

\item \code{library_linter}: Using library in package functions can have
unexpected side effects. (only for package functions)

\item \code{setwd_linter}: Changing the working directory in package functions
can have unexpected side effects. (only for package functions)

\item \code{source_linter}: Sourcing files in package functions can have
unexpected side effects. (only for package functions)
}}

