% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCodeStyle.R
\name{linterList}
\alias{linterList}
\title{List of linters tested to ensure INWT style conventions}
\usage{
linterList(type = c("script", "pkgFuns"))
}
\description{
Used in \code{\link{checkStyle}}. The set of included linters
depends on the type of the file. The following linters are always included:
\code{\link[lintr]{assignment_linter}},
\code{\link[lintr]{closed_curly_linter}},
\code{\link[lintr]{commas_linter}},
\code{\link{double_whitepace_linter}},
\code{\link[lintr]{infix_spaces_linter}},
\code{\link[lintr]{line_length_linter}},
\code{\link[lintr]{multiple_dots_linter}},
\code{\link[lintr]{no_tab_linter}},
\code{\link[lintr]{object_length_linter}},
\code{\link[lintr]{object_usage_linter}},
\code{\link[lintr]{open_curly_linter}},
\code{\link[lintr]{spaces_inside_linter}},
\code{\link[lintr]{spaces_left_parentheses_linter}},
\code{\link[lintr]{trailing_blank_lines_linter}},
\code{\link[lintr]{trailing_whitespace_linter}}
}
\examples{
# Code that lists all tested linters:
cat(paste0("\\\\code{\\\\link[lintr]{", sort(names(linterList())), "}}",
  collapse = ",\\n#' "))

}

