% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_formattingFigures.R
\name{makePct}
\alias{makePct}
\title{Create percentage from share}
\usage{
makePct(x, digits = 0, percSymbol = TRUE, keepTrailing0 = TRUE)
}
\arguments{
\item{x}{numeric: a share}

\item{digits}{integer: number of decimal places for the percentage}

\item{percSymbol}{logical: return the "\%" symbol?}

\item{keepTrailing0}{logical: keep trailing zeros (decimals) to have the same
length for all numbers?}
}
\value{
Character
}
\description{
Make a nice-looking percentage out of a share.
}
\examples{
makePct(0.25)
makePct(1.251, percSymbol = FALSE)
makePct(0.005, digits = 2)
makePct(0.26, digits = 2, keepTrailing0 = FALSE)
makePct(0.26, digits = 2, keepTrailing0 = TRUE)
}
