% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSkeleton.R
\name{projectSkeleton}
\alias{projectSkeleton}
\title{Create file structure}
\usage{
projectSkeleton(dir = ".", pkgName = NULL, pkgOnToplevel = TRUE,
  rProject = FALSE, exampleScript = TRUE, ...)
}
\arguments{
\item{dir}{Directory where the file structure is created, relative to the
current working directory. The current working directory by default.}

\item{pkgName}{character: If \code{pkgName} is specified, a package with this
name is created.}

\item{pkgOnToplevel}{logical: Should the package live in the main directory
(default) or in a subfolder called package?}

\item{rProject}{logical: Create an R Project?}

\item{exampleScript}{logical: Create example script?}

\item{...}{Further arguments passed to \code{\link[devtools]{create}} resp.
\code{\link[devtools]{setup}} if creating a package}
}
\description{
Creates a file structure with a useful folder structure. This
includes the folders data, reports and rScripts to store files. Furthermore,
the folders libWin and libLinux are created, together with an .Rprofile file.
Newly installed packages are then stored in the respective lib folder to
enable work in a sandbox. The infrastructure for a package and an RStudio
project can be added.
}
\examples{
projectSkeleton("tmp")

}
