% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_moreUtils.R
\name{read_csv}
\alias{read_csv}
\title{Read csv file checking for correct row number}
\usage{
read_csv(file, col_names = TRUE, skip = 0, n_max = Inf, ...)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector).

Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
be automatically uncompressed. Files starting with \code{http://},
\code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
downloaded. Remote gz files can also be automatically downloaded and
decompressed.

Literal data is most useful for examples and tests. It must contain at
least one new line to be recognised as data (instead of a path).}

\item{col_names}{Either \code{TRUE}, \code{FALSE} or a character vector
of column names.

If \code{TRUE}, the first row of the input will be used as the column
names, and will not be included in the data frame. If \code{FALSE}, column
names will be generated automatically: X1, X2, X3 etc.

If \code{col_names} is a character vector, the values will be used as the
names of the columns, and the first row of the input will be read into
the first row of the output data frame.

Missing (\code{NA}) column names will generate a warning, and be filled
in with dummy names \code{X1}, \code{X2} etc. Duplicate column names
will generate a warning and be made unique with a numeric prefix.}

\item{skip}{Number of lines to skip before reading data.}

\item{n_max}{Maximum number of records to read.}

\item{...}{arguments passed to \code{\link[readr]{read_csv}}}
}
\description{
Sometimes \code{\link[readr]{read_csv}} stops reading in a csv
file after several lines without any message. This function is a wrapper
around \code{\link[readr]{read_csv}} that counts the number of rows in the
original file and compares it with the resulting data.frame. If the numbers
of rows differ, it throws a warning.
}
\examples{
write.csv(x = iris, file = "iris.csv")
read_csv(file = "iris.csv")
read_csv(file = "iris.csv", skip = 3, col_names = FALSE, n_max = 10)

}
