% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_manipulateFilepath.R
\name{rmBackslash}
\alias{rmBackslash}
\title{Remove trailing "/" from path}
\usage{
rmBackslash(path)
}
\arguments{
\item{path}{charachter: A filepath}
}
\description{
Remove trailing "/" from path
}
