% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_rmPkgs.R
\name{rmPkgs}
\alias{rmPkgs}
\title{Remove non-standard packages from searchpath}
\usage{
rmPkgs()
}
\description{
Removes all non-standard packages from the searchpath. Only
packages loaded when starting R remain in the searchpath. This leads to a
clean searchpath without unwanted masking effects.
}
\examples{
\dontrun{
library(INWUtils)
search()
rmPkgs()
search()
# INWUtils is not in the searchpath anymore
}
}
