% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_createProjectSkeleton.R
\name{useSandbox}
\alias{useSandbox}
\title{Create a sandbox environment for package installation}
\usage{
useSandbox(dir)
}
\arguments{
\item{dir}{Project directory}
}
\description{
The following steps are taken:
\itemize{
  \item Create folders libWin, libLinux, libMac
  \item Write .gitignore into these folders so they can be commited without
  any real content
  \item Write .RProfile telling R to install and look for packages in libWin
  resp. libLinux by default
 }
}
