% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_checkStyle.R
\name{checkStyle}
\alias{checkStyle}
\title{Check code style}
\usage{
checkStyle(files, linters = selectLntrs(...), ...)
}
\arguments{
\item{files}{character vector: One or more filepaths}

\item{linters}{list: Named list of used linter functions}

\item{...}{Arguments passed to \code{\link{selectLntrs}}}
}
\description{
Checks a file for violations of the INWT style conventions using
\code{\link[lintr]{lint}}.
The \code{\link[lintr]{linters}} are partly taken from the package
\code{lintr}, and partly from this package.
}
\details{
Per default, the used linters are selected via
\code{\link{selectLntrs}}. If you pass a list of linters directly via the
\code{linters} argument, all arguments passed to \code{\link{selectLntrs}}
via \code{...} will be ignored.
}
\examples{
\dontrun{
writeLines(con = "lintExample.txt",
           text = c("# Example script to demonstrate INWT's own linters",
           # nolint start
                    "",
                    "foo <- function(x = 1, y) {",
                    "  2*x + 1",
                    "}",
                    "",
                    paste0("# This  line containts  double spaces and is ",
                           "very long. The following lines will use = ",
                           "instead of <- and access an internal INWT ",
                           "function."),
                    "z = 1",
                    "print(INWTUtils:::scriptLntrs())",
                    ""))
           # nolint end
checkStyle("lintExample.txt", type = "script")
unlink("lintExample.txt") # Remove file
}

}
