% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCodeStyle.R
\name{checkStyle}
\alias{checkStyle}
\title{Check code style}
\usage{
checkStyle(file, type = c("script", "pkgFuns"))
}
\arguments{
\item{file}{character: File to check}

\item{type}{character: Type of the file (script or pkgFuns)}
}
\description{
A file is checked for violations of the INWT style conventions
using \code{\link[lintr]{lint}}. In addition to the
\code{\link[lintr]{linters}} provided by the package \code{lintr}, some
custom linters are tested. For details about the tested linters see
\code{\link{selectLntrs}}. The set of used linters depends on the type of the
file. If the type is specified ("script" or "pkgFuns"), some additional
linters are tested.
}
\examples{
\dontrun{
# Write example file:
writeLines(c(
  "# This is an example document violating style conventions",
  "foo <- function() {",
  "library(INWTUtils)",
  "c(1+1 ,1) ",
  paste("print('A very long text which is nevertheless written into a",
  "single line such that the line exceeds 100 character by far')}"
)),
con = "example.R")

# Check file:
checkStyle(file = "example.R", type = "pkgFuns")
}

}

