% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCodeStyle.R
\name{checkStyle}
\alias{checkStyle}
\title{Check code style}
\usage{
checkStyle(file, type = c("script", "pkgFuns"))
}
\arguments{
\item{file}{character: File to check}

\item{type}{character: Type of the file (script or pkgFuns)}
}
\description{
A file is checked for violations of the INWT style conventions.
All types of files are checked using \code{\link[lintr]{lint}}. For details
about the tested linters see \code{\link{linterList}}. Further checks are
conducted if the type of the file is specified ("script" or "pkgFuns"). Those
checks are applied only to uncommented lines.
}
\examples{
\dontrun{
# Write example file:
writeLines(c(
  "# This is an example document violating style conventions",
  "foo <- function() {",
  "library(INWTUtils)",
  "c(1+1 ,1)  ",
  paste("print('A very long text which is nevertheless written into a",
  "single line such that the line exceeds 100 character by far'))}"
),
con = "example.R")

# Check file:
checkStyle(file = "example.R", type = "pkgFuns")
}

}

