% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCodeStyle.R
\name{checkText}
\alias{checkText}
\title{Check text file for a patterm}
\usage{
checkText(file, pattern)
}
\arguments{
\item{file}{character: Path to a file}

\item{pattern}{character: Pattern to look for}
}
\value{
If the text contains the pattern, a warning is returned, if not,
nothing is returned.
}
\description{
Checks a text file for a specific pattern. Regular expressions are allowed.
Only uncommented lines (i.e., lines which do not start with \code{#}) are
examined.
}
\examples{
\dontrun{
# Write example file:
writeLines(c(
  "# This is an example document to demonstrate the function checkTest",
  "x <- 1:3",
  "cat(INWTUtils:::forbiddenTextScript())"),
  con = "example.R")
  
# Check file:
checkText("example.R", ":::")
}

}

