% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSkeleton.R
\name{createPackage}
\alias{createPackage}
\title{Create package}
\usage{
createPackage(pkgName, pkgOnToplevel, dir, ...)
}
\arguments{
\item{pkgName}{character: Package name}

\item{pkgOnToplevel}{logical: Should the package live in the main
directory or in a subfolder called package?}

\item{dir}{character: Directory}

\item{...}{Further arguments passed to \code{\link[devtools]{create}} resp.
\code{\link[devtools]{setup}}}
}
\description{
Creates a package, either directly in the passed directory or
in a subfolder called package. Also creates infrastructure for testthat.
Internally used by \code{\link{projectSkeleton}}.
}

