% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSkeleton.R
\name{createPackage}
\alias{createPackage}
\title{Create package}
\usage{
createPackage(dir, pkgName, pkgOnToplevel, ...)
}
\arguments{
\item{dir}{character: Directory}

\item{pkgName}{character: Package name}

\item{pkgOnToplevel}{logical: Should the package live in the main
directory or in a subfolder called package?}

\item{...}{Further arguments passed to \code{\link[devtools]{create}} resp.
\code{\link[devtools]{setup}}}
}
\description{
Creates a package, either directly in the passed directory or
in a subfolder called package. Also creates an infrastructure for testthat, a
test of the package style. and an .Rbuildignore. An R project has to be
created separately.
Internally used by \code{\link{projectSkeleton}}.
}
\examples{
\dontrun{
createPackage(dir = "./", pkgName = "aTestPackage", pkgOnToplevel = FALSE)
dir.create("tmp")
createPackage(dir = "tmp/", pkgName = "aTestPackage", pkgOnToplevel = TRUE)
}

}

