% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSkeleton.R
\name{createProject}
\alias{createProject}
\title{Create R project}
\usage{
createProject(dir, pkg, pkgOnToplevel)
}
\arguments{
\item{dir}{character: Directory where the R project is created}

\item{pkg}{logical: Does the project contain a package?}

\item{pkgOnToplevel}{logical: Does the package live in the project directory
or a subfolder "package"?}
}
\description{
Creates an R project with useful configuration. If the project
contains a package, the respective options are set. Internally used by
\code{\link{projectSkeleton}}.
}

