% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmPkgs.R
\name{rmPkgs}
\alias{rmPkgs}
\title{Remove non-standard packages}
\usage{
rmPkgs()
}
\description{
Removes all non-standard packages from the searchpath. Only
packages loaded when starting R remain in the searchpath. This leads to a
clean searchpath without unwanted masking effects.
}
\examples{
\dontrun{
library(INWTutils)
search()
rmPkgs()
search()
# INWTutils is not in the searchpath anymore
}
}
